Converted lens.Lens to use van Laarhoven Lenses
lens.Lens now takes a function as it's only parameter. In haskell terms
this function has type `(a -> f a) -> b -> f b`. It's slightly more
difficult to construct lenses (internally, the api hasn't changed) now,
but it was a good learning exercise.

This change required adding a `lens.functor` module with Identity and
Const functors. Hopefully, I can resist adding a monad implementation
for a good while yet...