Reorganised lens definitions
The `lenses.lens` module will contain many default lenses and these will
all be exported to the root of the lenses package for easy access. Makes
for a much more clear api than having them defined on the `Lens` class
since it's not apparent if methods there will just return a brand new
lens or refine the focus of the current one.

The `Lens` class itself is the location to put various lens combinators
for manipulating lenses.