Added new lenses: item and item_by_value for dict manipualtion
item takes a key and focuses a tuple of the key and its value in the
dictionary. item_by_value focuses a key/value pair given a value. They
locate items using object identity, so be careful there. Also, you
cannot reasonably use item_by_value on a dictionary that has duplicates
of the value you're focusing on.