Allow composing lenses.Lens with & (`__and__`)
We need an easier to use syntax since lenses are composed a lot when
building them manually. Considered using `__add__`, but `&` is just as
convenient as `+` and we're far less likely to be doing bit-wise
operations on lenses any time soon. We're never going to use `__and__`
for it's intended meaning. Can always add `__add__` later if necessary.