added support for traversing empty sequences
This is done through the use of "functorisors" (terrible name), which
can pass around information about how to create pure versions of the
functors. Still doesn't work for `BaseLens.get` since we also need to
know the `mempty` of the type the Const is supposed to contain, but it
will have to do for now. `get_all`, `set`, and `modify` all work fine.